Refactor comments logic out of App into Discussion
