+++ index.d.ts
@@ -652,7 +652,8 @@ type IslandType =
     | 'sub-nav-root'
     | 'edition-root'
     | 'most-viewed-right'
-    | 'share-comment-counts'
+    | 'share-count-root'
+    | 'comment-count-root'
     | 'most-viewed-footer'
     | 'reader-revenue-links-footer'
     | 'slot-body-end'
