+++ src/web/components/App.tsx
@@ -3,8 +3,8 @@ import { useAB } from '@guardian/ab-react';
 import { tests } from '@frontend/web/experiments/ab-tests';
 
 import { EditionDropdown } from '@frontend/web/components/EditionDropdown';
+import { ShareCount } from '@frontend/web/components/ShareCount';
 import { MostViewedFooter } from '@frontend/web/components/MostViewed/MostViewedFooter/MostViewedFooter';
-import { Counts } from '@frontend/web/components/Counts';
 import { RichLinkComponent } from '@frontend/web/components/elements/RichLinkComponent';
 import { CalloutBlockComponent } from '@root/src/web/components/elements/CalloutBlockComponent';
 import { YoutubeBlockComponent } from '@root/src/web/components/elements/YoutubeBlockComponent';
@@ -13,7 +13,7 @@ import { SlotBodyEnd } from '@frontend/web/components/SlotBodyEnd';
 import { Links } from '@frontend/web/components/Links';
 import { SubNav } from '@frontend/web/components/SubNav/SubNav';
 import { GetMatchNav } from '@frontend/web/components/GetMatchNav';
-import { CommentsLayout } from '@frontend/web/components/CommentsLayout';
+import { Discussion } from '@frontend/web/components/Discussion';
 import { StickyBottomBanner } from '@root/src/web/components/StickyBottomBanner/StickyBottomBanner';
 import { SignInGateSelector } from '@root/src/web/components/SignInGate/SignInGateSelector';
 
@@ -39,10 +39,8 @@ import { toTypesPillar } from '@root/src/lib/format';
 import { initPerf } from '@root/src/web/browser/initPerf';
 import { getCookie } from '@root/src/web/browser/cookie';
 import { getCountryCode } from '@frontend/web/lib/getCountryCode';
-import { getDiscussion } from '@root/src/web/lib/getDiscussion';
 import { getUser } from '@root/src/web/lib/getUser';
 import { getBrazeUuid } from '@root/src/web/lib/getBrazeUuid';
-import { getCommentContext } from '@root/src/web/lib/getCommentContext';
 import { FocusStyleManager } from '@guardian/src-foundations/utils';
 import { incrementAlreadyVisited } from '@root/src/web/lib/alreadyVisited';
 import { incrementDailyArticleCount } from '@frontend/web/lib/dailyArticleCount';
@@ -109,19 +107,6 @@ const GetMatchStats = React.lazy(() => {
 
 type Props = { CAPI: CAPIBrowserType; NAV: NavType };
 
-const commentIdFromUrl = () => {
-    const { hash } = window.location;
-    if (!hash) return;
-    if (!hash.includes('comment')) return;
-    if (!hash.split('-')[1]) return;
-    return parseInt(hash.split('-')[1], 10);
-};
-
-const hasCommentsHashInUrl = () => {
-    const { hash } = window.location;
-    return hash && hash === '#comments';
-};
-
 const componentEventHandler = (
     componentType: any,
     id: any,
@@ -154,23 +139,8 @@ export const App = ({ CAPI, NAV }: Props) => {
     // banners need countryCode but we don't want to block all banners from
     // executing their canShow logic until countryCode is available):
     const [asyncCountryCode, setAsyncCountryCode] = useState<Promise<string>>();
-    const [commentCount, setCommentCount] = useState<number>(0);
-    const [isClosedForComments, setIsClosedForComments] = useState<boolean>(
-        true,
-    );
-    const [commentPage, setCommentPage] = useState<number>();
-    const [commentPageSize, setCommentPageSize] = useState<25 | 50 | 100>();
-    const [commentOrderBy, setCommentOrderBy] = useState<
-        'newest' | 'oldest' | 'recommendations'
-    >();
-    const [isExpanded, setIsExpanded] = useState<boolean>(false);
-    const [hashCommentId, setHashCommentId] = useState<number | undefined>(
-        commentIdFromUrl(),
-    );
 
     const [shouldUseAcast, setShouldUseAcast] = useState<boolean>(false);
-
-    const hasCommentsHash = hasCommentsHashInUrl();
 
     const pageViewId = window.guardian?.config?.ophan?.pageViewId;
 
@@ -226,29 +196,6 @@ export const App = ({ CAPI, NAV }: Props) => {
     }, []);
 
     useEffect(() => {
-        const callFetch = async () => {
-            const response = await getDiscussion(
-                CAPI.config.discussionApiUrl,
-                CAPI.config.shortUrlId,
-            );
-            setCommentCount(
-                (response && response.discussion.commentCount) || 0,
-            );
-            setIsClosedForComments(
-                response && response.discussion.isClosedForComments,
-            );
-        };
-
-        if (CAPI.isCommentable) {
-            callFetch();
-        }
-    }, [
-        CAPI.config.discussionApiUrl,
-        CAPI.config.shortUrlId,
-        CAPI.isCommentable,
-    ]);
-
-    useEffect(() => {
         incrementAlreadyVisited();
     }, []);
 
@@ -265,28 +212,6 @@ export const App = ({ CAPI, NAV }: Props) => {
         };
         incrementArticleCountsIfConsented();
     }, []);
-
-    // Check the url to see if there is a comment hash, e.g. ...crisis#comment-139113120
-    // If so, make a call to get the context of this comment so we know what page it is
-    // on.
-    useEffect(() => {
-        if (hashCommentId) {
-            getCommentContext(CAPI.config.discussionApiUrl, hashCommentId).then(
-                (context) => {
-                    setCommentPage(context.page);
-                    setCommentPageSize(context.pageSize);
-                    setCommentOrderBy(context.orderBy);
-                    setIsExpanded(true);
-                },
-            );
-        }
-    }, [CAPI.config.discussionApiUrl, hashCommentId]);
-
-    useEffect(() => {
-        if (hasCommentsHash) {
-            setIsExpanded(true);
-        }
-    }, [hasCommentsHash]);
 
     // Ensure the focus state of any buttons/inputs in any of the Source
     // components are only applied when navigating via keyboard.
@@ -407,16 +332,6 @@ export const App = ({ CAPI, NAV }: Props) => {
     const display: Display = decideDisplay(CAPI);
     const adTargeting: AdTargeting = buildAdTargeting(CAPI.config);
 
-    const handlePermalink = (commentId: number) => {
-        window.location.hash = `#comment-${commentId}`;
-        const comment = window.document.getElementById(`comment-${commentId}`);
-        if (comment) {
-            // The comment was already on the page so just scroll to it.
-            comment.scrollIntoView();
-        }
-        setHashCommentId(commentId);
-        return false;
-    };
     return (
         // Do you need to Hydrate or do you want a Portal?
         //
@@ -447,6 +362,12 @@ export const App = ({ CAPI, NAV }: Props) => {
                     dataLinkName="nav2 : topbar : edition-picker: toggle"
                 />
             </Hydrate>
+            <Portal root="share-count-root">
+                <ShareCount 
+                    ajaxUrl={CAPI.config.ajaxUrl}
+                    pageId={CAPI.pageId}
+                />
+            </Portal>
             {CAPI.youtubeMainMediaBlockElement.map((youtubeBlock, index) => (
                 <Hydrate
                     key={index}
@@ -653,24 +574,6 @@ export const App = ({ CAPI, NAV }: Props) => {
                     />
                 </Hydrate>
             ))}
-            <Portal root="share-comment-counts">
-                {CAPI.isCommentable ? (
-                    <Counts
-                        ajaxUrl={CAPI.config.ajaxUrl}
-                        pageId={CAPI.config.pageId}
-                        commentCount={commentCount}
-                        pillar={pillar}
-                        setIsExpanded={setIsExpanded}
-                    />
-                ) : (
-                    <Counts
-                        ajaxUrl={CAPI.config.ajaxUrl}
-                        pageId={CAPI.config.pageId}
-                        pillar={pillar}
-                        setIsExpanded={setIsExpanded}
-                    />
-                )}
-            </Portal>
             <Portal root="most-viewed-right">
                 <Lazy margin={100}>
                     <Suspense fallback={<></>}>
@@ -748,29 +651,20 @@ export const App = ({ CAPI, NAV }: Props) => {
             <Portal root="sign-in-gate">
                 <SignInGateSelector isSignedIn={isSignedIn} CAPI={CAPI} />
             </Portal>
-
-            {/* Don't lazy render comments if we have a comment id in the url or the comments hash. In
-                these cases we will be scrolling to comments and want them loaded */}
             <Hydrate root="comments">
-                <CommentsLayout
-                    user={user}
+                <Discussion
+                    discussionApiUrl={CAPI.config.discussionApiUrl}
+                    shortUrlId={CAPI.config.shortUrlId}
+                    isCommentable={CAPI.isCommentable}
                     pillar={pillar}
-                    baseUrl={CAPI.config.discussionApiUrl}
-                    shortUrl={CAPI.config.shortUrlId}
-                    commentCount={commentCount}
-                    commentPage={commentPage}
-                    commentPageSize={commentPageSize}
-                    commentOrderBy={commentOrderBy}
-                    isClosedForComments={isClosedForComments}
+                    user={user}
                     discussionD2Uid={CAPI.config.discussionD2Uid}
                     discussionApiClientHeader={
                         CAPI.config.discussionApiClientHeader
                     }
                     enableDiscussionSwitch={CAPI.config.enableDiscussionSwitch}
-                    expanded={isExpanded}
-                    commentToScrollTo={hashCommentId}
-                    onPermalinkClick={handlePermalink}
-                    hideAds={CAPI.isAdFreeUser || CAPI.shouldHideAds}
+                    isAdFreeUser={CAPI.isAdFreeUser}
+                    shouldHideAds={CAPI.shouldHideAds}
                 />
             </Hydrate>
             <Portal root="most-viewed-footer">
